refactor users restfull api
