+++ tests/test_api_change_password.py
@@ -21,7 +21,7 @@ class FlaskTest(unittest.TestCase):
                     data=VALID_PROFILE_ADMIN,
                     content_type='application/json')
 
-        tester.post("/register/",
+        tester.post("/user/",
                     data=VALID_ADMIN1_REGISTER,
                     content_type='application/json')
 
@@ -54,7 +54,7 @@ class FlaskTest(unittest.TestCase):
                     data=VALID_PROFILE_ADMIN,
                     content_type='application/json')
 
-        response_user = tester.post("/register/",
+        response_user = tester.post("/user/",
                                     data=json.dumps({'first_name': 'Gonza', 'last_name': 'Paez', 'email':
                                         'algo@algo.com', 'password': old_pass, 'national_id': '12345678',
                                                      'national_id_type': 'DNI', 'alias': 'gonzalgo', 'profile': 0}),
@@ -92,7 +92,7 @@ class FlaskTest(unittest.TestCase):
                     data=VALID_PROFILE_ADMIN,
                     content_type='application/json')
 
-        response_user = tester.post("/register/",
+        response_user = tester.post("/user/",
                                     data=json.dumps({'first_name': 'Gonza', 'last_name': 'Paez', 'email':
                                         'algo@algo.com', 'password': old_pass, 'national_id': '12345678',
                                                      'national_id_type': 'DNI', 'alias': 'gonzalgo', 'profile': 0}),
