+++ tests/test_api_login.py
@@ -21,7 +21,7 @@ class FlaskTest(unittest.TestCase):
                     data=VALID_PROFILE_ADMIN,
                     content_type='application/json')
 
-        tester.post("/register/",
+        tester.post("/user/",
                     data=VALID_ADMIN1_REGISTER,
                     content_type='application/json')
 
@@ -41,7 +41,7 @@ class FlaskTest(unittest.TestCase):
                     data=VALID_PROFILE_ADMIN,
                     content_type='application/json')
 
-        tester.post("/register/",
+        tester.post("/user/",
                     data=VALID_ADMIN1_REGISTER,
                     content_type='application/json')
 
@@ -75,7 +75,7 @@ class FlaskTest(unittest.TestCase):
                     data=VALID_PROFILE_ADMIN,
                     content_type='application/json')
 
-        tester.post("/register/",
+        tester.post("/user/",
                     data=VALID_ADMIN1_REGISTER,
                     content_type='application/json')
